+++ examples/tokio_send.rs
@@ -33,6 +33,6 @@ async fn main() -> Result<()> {
         socket_tx.write_frame(frame).await?;
 
         println!("Waiting 3 seconds");
-        Delay::new(Duration::from_secs(3)).await?;
+        Delay::new(Duration::from_secs(3)).await;
     }
 }